src/azureFunction/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai
azure-search-documents
azure.identity
llama-index
llama-index-vector-stores-azureaisearch
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
llama-index-readers-azstorage_blob
azure.identity